+++ src/assets/css/globals.css
@@ -63,9 +63,14 @@ https://fonts.coollabs.io/ A privacy-friendly drop-in replacement for Google Fon
     background-color: theme(colors.accent.DEFAULT / 50%);
   }
 
-  body {
+  body, html {
+    width: 100%;
+    height: 100%;
+    overflow: hidden;
+    margin: 0;
     font-family: 'Noto Serif', serif;
     font-weight: 400;
+    font-size: calc(50px * (100vh / 1080));
   }
 
   h1, h2, h3, h4, h5 {
