+++ src/pages/About/index.tsx
@@ -1,0 +1,12 @@
+import AboutMeSection from "./AboutMeSection";
+import RecentProjectsSection from "./RecentProjectsSection";
+
+function About() {
+  return (
+    <div>
+        <AboutMeSection />
+        <RecentProjectsSection/>
+    </div>
+  )
+}
+export default About;
