+++ src/routes/layouts/main.tsx
@@ -5,7 +5,7 @@ import { Outlet } from "react-router-dom";
 
 export default function Layout() {
   return (
-    <div className="bg-background text-foreground flex h-screen min-h-screen w-screen flex-col justify-between overflow-y-scroll lg:snap-y">
+    <div className="min-w-[1024px] bg-background text-foreground flex h-screen min-h-screen w-screen flex-col justify-between overflow-y-scroll lg:snap-y">
       <Navbar />
       <main className="container grow">
         <Outlet />
